// src/services/news.service.d.ts
export declare const newsService: {
    getAllNews(): Promise<{
        id: string;
        title: string;
        slug: string;
        category: string;
        categoryColor: string;
        excerpt: string;
        content: string;
        image: string;
        date: string;
        readTime: string;
        author: string;
        featured: boolean;
        createdAt: Date;
        updatedAt: Date;
    }[]>;
    getNewsById(id: string): Promise<{
        id: string;
        title: string;
        slug: string;
        category: string;
        categoryColor: string;
        excerpt: string;
        content: string;
        image: string;
        date: string;
        readTime: string;
        author: string;
        featured: boolean;
        createdAt: Date;
        updatedAt: Date;
    } | null>;
    getNewsBySlug(slug: string): Promise<{
        id: string;
        title: string;
        slug: string;
        category: string;
        categoryColor: string;
        excerpt: string;
        content: string;
        image: string;
        date: string;
        readTime: string;
        author: string;
        featured: boolean;
        createdAt: Date;
        updatedAt: Date;
    } | null>;
    createNews(data: {
        title: string;
        slug: string;
        category: string;
        categoryColor: string;
        excerpt: string;
        content: string;
        image: string;
        date: string;
        readTime: string;
        author: string;
        featured: boolean;
    }): Promise<{
        id: string;
        title: string;
        slug: string;
        category: string;
        categoryColor: string;
        excerpt: string;
        content: string;
        image: string;
        date: string;
        readTime: string;
        author: string;
        featured: boolean;
        createdAt: Date;
        updatedAt: Date;
    }>;
    updateNews(id: string, data: any): Promise<{
        id: string;
        title: string;
        slug: string;
        category: string;
        categoryColor: string;
        excerpt: string;
        content: string;
        image: string;
        date: string;
        readTime: string;
        author: string;
        featured: boolean;
        createdAt: Date;
        updatedAt: Date;
    }>;
    deleteNews(id: string): Promise<{
        id: string;
        title: string;
        slug: string;
        category: string;
        categoryColor: string;
        excerpt: string;
        content: string;
        image: string;
        date: string;
        readTime: string;
        author: string;
        featured: boolean;
        createdAt: Date;
        updatedAt: Date;
    }>;
};
//# sourceMappingURL=news.service.d.ts.map